requirements.txt
# Core LangChain library for building agent applications
langchain>=0.1.0

# OpenAI integration for LangChain
langchain-openai>=0.0.5

# Python library for loading environment variables from .env files
python-dotenv>=1.0.0

# OpenAI Python client (dependency for langchain-openai)
openai>=1.0.0
